Working proof of concept
works! Well sort of... It's working where the robot finds the best path and then lays down obstacles to force the human onto it. Still need to add checks every time step to make sure the human is adhering to the plan and to run the optimizer across multiple paths to find the optimal one.